+++ src/assets/index.ts
@@ -1,2 +1,3 @@
 export * from "./intro";
 export * from "./loading";
+export * from "./outtro";
