+++ src/assets/outro/index.ts
@@ -1,0 +1,1 @@
+export { default as Outro } from "./outro.png";
