--- src/assets/outtro/index.ts
@@ -1,1 +1,0 @@
-export { default as Outtro } from "./outtro.png";
